requirements.txt
certifi==2020.12.5
cloudpickle==1.6.0
colour==0.1.5
future==0.18.2
gym==0.18.0
numpy==1.20.2
Pillow==7.2.0
pygame==2.0.0
pyglet==1.5.0
python-twitch-stream==1.0.2
scipy==1.6.3
